requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
openai>=1.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
msal>=1.20.0
pydantic>=1.10.0
pytest>=7.0.0
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
